+++ android/src/main/kotlin/com/example/nfc_classic_mifare/AccessType.kt
@@ -1,0 +1,5 @@
+package com.example.nfc_classic_mifare
+
+enum class AccessType {
+    Never, KeyA, KeyB, KeyAB
+}