+++ android/src/main/kotlin/com/example/nfc_classic_mifare/SectorAccessConditions.kt
@@ -1,0 +1,12 @@
+package com.example.nfc_classic_mifare
+
+class SectorAccessConditions(
+    val block0: DataBlockAccessConditions,
+    val block1: DataBlockAccessConditions,
+    val block2: DataBlockAccessConditions,
+    val sectorBlock: SectorBlockAccessConditions
+) {
+    object Common {
+        val NEVER = Utils.resolveSectorAccessConditions()
+    }
+}