+++ example/android/app/src/main/kotlin/com/example/mifare_nfc_classic_example/MainActivity.kt
@@ -1,21 +1,6 @@
-package com.example.mifare_nfc_classic_example
+package com.example.nfc_classic_mifare_example
 
-import android.app.PendingIntent
-import android.content.Intent
-import android.nfc.NfcAdapter
 import io.flutter.embedding.android.FlutterActivity
 
-
-class MainActivity : FlutterActivity() {
-    override fun onResume() {
-        super.onResume()
-        val intent = Intent(context, javaClass).addFlags(Intent.FLAG_ACTIVITY_SINGLE_TOP)
-        val pendingIntent = PendingIntent.getActivity(context, 0, intent, 0)
-        NfcAdapter.getDefaultAdapter(context)?.enableForegroundDispatch(this, pendingIntent, null, null)
-    }
-    override fun onPause() {
-        super.onPause()
-        NfcAdapter.getDefaultAdapter(context)?.disableForegroundDispatch(this)
-    }
-
+class MainActivity: FlutterActivity() {
 }