+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## 0.3.5
+
+- Package name change to nfc_mifare_classic to be able to publish
+
+## 0.3.1
+
+- Thanks frost768
+  - access bits decoding for read and write operations
+  - keyA and keyB arguments added
+  - library api updated
+
+## 0.3.0
+
+- Implemented null safety
+
 ## 0.2.1
 
 - Added function to overwrite a block